pyproject.toml
[tool.poetry]
name = "rlt-telegram-bot"
version = "0.1.0"
description = ""
authors = ["bulatsabirow <bulatsabirow@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
motor = "^3.4.0"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.1.1"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
